+++ docs/ChangeLog.md
@@ -5,6 +5,7 @@
 ## Upcoming
 
 - **Feature**: Ability to export a return as an excel file
+- **Update**: Don't add sample fixtures when adding a new MCA
 
 ## 7th July 2025